+++ templates/subscribe5.md.js
@@ -1,0 +1,8 @@
+md(`
+${raster("Spiritual Maintenance book by Aaron Brewer")}
+
+##### Spiritual Maintenance
+${icons({
+    'amazon': 'https://a.co/d/0doSi5ML',
+})}
+`)